138548 by corwin on 2019/03/28 (pair: msakai) Merged 137997 to linux-vdo. 137997 by corwin on 2019/02/11 (pair: msakai) Converted BlockMap to support suspend and resume [VDO-4402].
This change implements suspend and resume semantics for the block map
via new drainBlockMap() and resumeBlockMap() methods. With this change,
there are three flavors of drain: flush, suspend, and save. Both flush
and save trigger a write of all drity block map metadata (both the
trees and page cache). Both suspend and save leave the block map
quiescent, requiring a resume before the block map may be used again by
DataVIOs.

By merging the various flavors of drain into a single method,
leveraging the change to make the page cache aware of the block map,
eliminating the explicit drain of the trees' object pools, and allowing
the drain of the trees to trigger the drain of the page cache, the
number of completions and callbacks required to drain the block map has
been reduced.

.../blockMap.h
.../blockMapInternals.h
.../blockMap.c
Merged flushBlockMap() and closeBlockMap() into drainBlockMap() and
reworked to leverage the new infrastructure (ActionManager,
AdminState) and new component awareness to simplify the
implementation. Added an AdminState field to BlockMap and removed
the BlockMap's now unused zone completion.
Added resumeBlockMap().

.../completion.h
.../completion.c
Removed BLOCK_MAP_ZONE_COMPLETION.

.../vdoPageCache.h
.../vdoPageCacheInternals.h
.../vdoPageCache.c
Renamed flushVDOPageCacheAsync() to drainVDOPageCache() and added
support for suspending without flushing all dirty pages.
Removed the check for busy pages from drainVDOPageCache() since by
contract, the block map should only be suspended after all
DataVIOs have been drained. Furthermore, any outstanding busy
pages will correctly delay the drain operation until they are
returned, so there is no harm in just waiting for them
(vs. potential use-after-free errors if we don't).
Added assertIOAllowed() to check whether the cache is not quiescent and
added calls to it before anywhere the cache will issue I/O.
Removed the 'flushCompletion' field now that the completion waiting for
the page cache to drain is stored in the cache's BlockMapZone's
AdminState.
Removed syncVDOPageCacheAsync() which was only used for testing and
relies on mechanisms which have been removed. It is easy for the
unit test which used it to replace it with a call to the new drain
method followed by a resume.

.../blockMapTree.h
.../blockMapTree.c
Merged closeZoneTrees() and suspendZoneTrees() into drainZoneTrees().
Modified the drain process to check that the zone's VIO pool is not
busy instead of explicitly suspending the pool. This eliminates the
need for an extra completion.
Modified checkForIOComplete() to drain the page cache instead of
(implicitly) finishing the tree drain and relying on another level
of callbacks to drain the page cache.
Removed resumeZoneTrees() since the BlockMapZone's state is now managed
directly from resumeBlockMap() and the VIO pool is no longer suspended
when the trees are.

.../vdoResizeLogical.c
.../blockMap.h
.../blockMap.c
Moved the draining of the block map from growBlockMap() into
the resize transaction in vdoResizeLogical. Also added a block map
resume to the transaction. Once the VDO has a real suspend operation,
the suspend and resume in the resize operation will go away since
it will only be legal to run the resize operation while already
suspended.

.../blockMapRecovery.c
Reflected changes by using new drainBlockMap() instead of
flushVDOPageCacheAsync since flushing one page cache without
flushing the rest of the block map is no longer supported. This
requires making the recovery completion's sub-task completion
enqueueable since most of the transaction runs on logical zone
0, but drainBlockMap() must be called from the admin thread.

.../objectPool.h
.../objectPool.c
Added isPoolBusy() to allow users of a pool to check that the pool's
resources are not in use rather than suspending the pool. Once
the block allocator is also converted to the new semantics, the
suspend and resume methods on object pool can go away.

.../adminState.h
.../adminState.c
Added ADMIN_STATE_FLUSHING and associated predicate; modified
isDraining() and isDrainOperation() to be true for the flushing state.
Modified finishDraining() to handle the flushing state, which, when
finished, goes back to normal operation without requiring a resume.
Removed warn_unused_result from resumeIfQuiescent() since some of its
callers legitimately don't care about the return value.

.../referenceCountRebuild.c
.../vdoClose.c
Reflected changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>