138595 by corwin on 2019/04/01 (pair: msakai) Merged 138088, 138113, and 138114 to linux-vdo. 138088 by corwin on 2019/02/21 (pair: msakai) Added suspend and resume of the slab scrubber [VDO-4402].
As part of implementing suspend and resume semantics for the slab
depot, it is useful to be able to suspend and then resume the slab
scrubber. While not strictly necessary for the current bug, it affords
the ability to better handle some of the other growth scenarios which
were discussed in Brno (i.e. it is a step towards being able to grow
while in recovery mode).

This change also separates the completions used to notify the slab
depot when scrubbing is complete from the completion used to notify the
block allocator that scrubbing has been stopped. Doing so makes the
subsequent drain implementation of slab depot simpler.

.../slabScrubber.h
.../slabScrubberInternals.h
.../slabScrubber.c
Reworked the scrubbing loop to provide the ability to resume scrubbing
once it has been stopped. To do so, replaced the "isScrubbing" and
"stopScrubbing" fields with an AdminState. In addition to the new
functionality, did some refactoring and rearrangement.
Added a "parent" parameter to stopScrubbing(). This new completion is
notified when the scrubber is stopped. The completion supplied to
scrubSlabs() is now only finished when the scrubber has actually
finished scrubbing either due to an error or because all the slabs
have been processed. Separating these two completions affords a
simplification in BlockAllocator (described below).
Added resumeScrubbing().
Removed isScrubbing() since it now has no callers.
Removed the threadID parameter from scrubSlabs() since it is always
called on the allocator thread.

.../adminState.h
.../adminState.c
Added getAdminStateName().
Removed the warn_unused_result attribute from startDraining() as it now
has a caller which legitimately does not care about the return value.

.../blockAllocatorInternals.h
.../blockAllocator.c
Leveraged the changes to SlabScrubber to simplify the implementation
of closeBlockAllocator(), transforming stopping and waiting on the
scrubber into a new closeStep. Since the scrubber uses the slab
completion to save scrubbed slabs, this necessitated converting
doCloseAllocatorStep() to use the allocator completion instead of
the slab completion.
Removed scrubberFinished(). Now that the allocator uses the allocator
close step mechanism to stop the scrubber, the scrubbing complete
notification can call notifyZoneFinishedScrubbing() on the slab depot
directly. Also set the error handler for scrubbing to the new
noopCallback() method since notifyZoneFinishedScrubbing() now expects
to be called only when scrubbing was successful.

.../slabDepotInternals.h
.../slabDepot.c
Renamed notifyZoneStoppedScrubbing() to
notifyZoneFinishedScrubbing() since it is now only called when
scrubbing completes successfully, not when scrubbing is stopped.
Modfied the method to be a VDOAction which expects the slab
depot to be the completion's parent.  Removed the attempt to
preserve the scrubbing error both because it was redundant (the
error was already logged and read-only mode already entered at
the source of the error), and because it was not thread safe.

.../completion.h
Added noopCallback() to be used as an error handler in a case where
there's nothing to do.

138113 by corwin on 2019/02/25 (pair: msakai)
Fixed race in SlabScrubber introduced by change 138088 [VDO-4402].

.../slabScrubber.c
Fixed a bug in finishScrubbing() which would result in a spurious
VDO_INVALID_ADMIN_STATE error if the scrubber was told to stop
while it was already scrubbing the final slab to be scrubbed.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>