139941 by tomj on 2019/07/09 (pair: msakai) Merge change 139883 to Jasper 139883 by tomj on 2019/07/03 (pair: msakai) Remove the session module from linux-uds.
When it was created, the session module contained code that was shared
by the BlockContext and IndexSession.  But today the BlockContext is no
longer needed, and we have removed it.

The session ID mechanism also let us hide some UDS data structures from
our users.  This seemed important when we were proprietary software,
but now that we are open source it is far less important.

The usage of an ID instead of a simple structure pointer is not a
mechanism that the kernel uses.  Eliminating the session now will avoid
the need to defend the practice when we go upstream.

An effect of this change is that we no longer have a list of all the
open index sessions.  The primary use of this list was to search it
when we needed to map a session ID to a IndexSession pointer.  We no
longer need to do this search, and get the additional benefit of not
needing a mutex to protect the list while we search it.

A downside of not having the list of open index sessions is that we
used this list when we wanted to close every index session.  We would
shut down an albserver by sending it a ^C, and the exit handler that
UDS set up would be able to close the open index.  In the kernel, we do
not have a similar mechanism.  The UDS user must explicitly close the
index before the uds module is removed.  If we find a production UDS
user mode customer, we will need to decide whether to recreate a list
of open index sessions to use for this purpose.

This change will require followups that do a lot of cleanup.  There are
multiple XXX comments left in the code to identify those cleanups.

.../uds.h
Dispense with using an opaque number to identify an IndexSession, and
just use a "struct indexSession *".

.../indexSession.h
.../indexSession.c
Delete the index session group, which is no longer needed.
Refactor the request count and its associated mutex and condition
variable out of the Session and into the IndexSession.
Remove the conversion of a session ID into an IndexSession from the
getIndexSession method.  Retain the checking of the IndexSession
status and incrementing the IndexSession request count.
Create the new waitForNoRequestsInProgress method, which replaces the
old usage of the waitForIdleSession method.

.../request.h
Add an XXX comment about the duplicate copy of the IndexSession
pointer.

.../request.c
Use the modified getIndexSession method.

.../udsMain.c
Remove all code related to the index session group from the
makeIndexSession method.

.../udsState.h
.../udsState.c
Remove the code that closes all the open index sessions in the process
exit code, and create an XXX comment about it.
Remove the index session group and its associated methods.
Remove the global methods that operate the global state mutex, which
is now used only internally by the udsState module.  It is no longer
used to protect the index session group.

.../session.h
.../session.c
Delete the session module.

139945 by tomj on 2019/07/09 (pair: msakai)
Remove support for the session module from UDS

.../Makefile
Do not ship or build the session module.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>