139950 by tomj on 2019/07/09 (pair: msakai) ALB-3015 Merge change 139891 to Jasper 139891 by tomj on 2019/07/05 (pair: sweettea) ALB-3015 Merge change 139882 to linux-uds 139882 by tomj on 2019/07/03 (pair: msakai) ALB-3015 Reduce the default number of index zones.
Our performance tests report that the performance improves when we
add zones up to the maximum number of cores.  But we run our tests
on ALBIREO-PMI or VDO-PMI hosts, which only have 6 cores.  In running
these tests using Beaker, we have gotten to see test results from hosts
with 16 cores, and see two observations.  One is that the performance
is limited by the test having only one thread doing hashing, so that
the measured performance levels off at the point it is limited by the
hash speed.  The other is that on our VDO-PMI hosts, we can end up
using most of the CPU power of the 6 cores.

On real systems, we cannot afford to let UDS use 100% of the CPU
capacity.  We need to leave cycles for VDO to do its work, and for the
application code that is using the VDO device.

This change implements using a default value for the UDS zones that is
only one-half of the number of cores.

.../zone.c
Set the default number of zones to be one-half of the number of cores.
Keep using the code that enforces a minimum of 1 zone and a maximum
of 16 zones.

139955 by corwin on 2019/07/09 (pair: msakai)
Merged 138317 to linux-vdo.
138317 by corwin on 2019/03/15 (pair: msakai)
Improved ActionManager [VDO-4402].

This change makes several improvements to ActionManager which help to
avoid potential use-after-free races when ActionManagers are used to run
actions which, when complete, send notifications which can result in
freeing the ActionManager. It also removes unused ActionManager
functionality and leverages that removal to improve the implementation
of ActionManager.

.../actionManager.h
.../actionManager.c
Added an optional preamble to actions. The preamble is run on the
initiator thread before applying the action to each zone.
Renamed ActionConclusion to InitiatorAction and changed the return type
from void to int. InitiatorAction is the type of both the preamble and
conclusion. The return type allows these method to return an error
code if they fail. If the preamble fails, the action will not be
applied to the zones, but the conclusion will be run.
Added an ActionScheduler type and scheduler constructor parameter and
field. This method, if supplied is used to try scheduling a next
action whenever an action completes. This method simplifies the use of
ActionManagers for the BlockMap and SlabDepot which need to try an era
advance or tail commit after other actions. By adding the scheduler,
other conclusions don't need to call the common bottleneck to schedule
these repeating actions.
Removed continueAction() as it was unused.
Removed hasNextAction(), getCurrentActionStatus() as none of them has
any callers as a result of this change.
Renamed the "action" field of Action and the "action" parameter of
scheduleAction() to "zoneAction" which is less confusing and awkward.
Leveraged the removal of continueAction() to simplify the managing of
action slots. The Action structure acquired 'inUse' and 'next' fields,
and the actions are managed via a ring of next pointers (though not
using RingNode). The 'nextAction' and 'acting' fields were removed
as they are no longer needed.
Modified finishActionCallback() to avoid potential use-after-free races
by ensuring that once the conclusion has been run and the parent of an
action has been notified, further accessing of the ActionManager only
happens in cases where the ActionManager can not have been freed as a
result of the notification.  This is predicated on no further actions
being scheduled if the current action is preparing for a shutdown,
although the ActionManager itself can not enforce this.

.../actionManager.c
.../completion.h
.../completion.c
Factored ActionManager's handleZoneError() into new
preserveErrorAndContinue() method in completion.[hc]. This idiom is
used in other places, and future changes will convert other code to
use it. It will also see more use in future suspend/resume changes.

.../blockMapInternals.h
.../blockMap.c
Reflected ActionManager changes:
Added a default scheduler for era advance. Removed the "aging" field.
Reworked resume and drain to avoid potential use-before-free errors
and improve error handling.
Removed the "previousEraPoint" field and both places where it is set as
this field had become write only.

.../slabDepotInternals.h
.../slabDepot.c
Reflected ActionManager changes:
Added a default scheduler for oldest tail commit. Removed the
"lockReleaseActive" field.
Reworked methods using the ActionManager to avoid potential
use-before-free errors and improve error handling.

Co-authored-by: tomj <tom@jaskiewicz.us>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: corwin <corwin@redhat.com>