139956 by corwin on 2019/07/09 (pair: msakai) Merged 138432 to linux-vdo. 138432 by corwin on 2019/03/22 (pair: bjohnston, msakai, sweettea) Unified slab depot drains into a single drain method [VDO-4402].
.../actionManager.h
.../actionManager.c
Modified ActionManager to have an AdminState and added an
AdminStateCode to the Action structure. This allows the scheduling
of actions which involve an AdminState change to only change the
state once the action is launched without having to store the
operation in the owner of the ActionManager.
Added scheduleOperation() which behaves just as scheduleAction() but
with an added AdminStateCode 'operation' parameter.
Modified scheduleAction() to become a wrapper for scheduleOperation()
with a default operation code.
Added getCurrentManagerOperation() to allow ZoneActions to read the
current operation from the action manager.

.../adminState.h
.../adminState.c
Broke out the various components of an AdminStateCode into separate
enums for clarity and documented what the various states mean.
Added startOperation() and finishOperation() for use by ActionManager.
These change the state but don't take a waiter.
Added setOperationWaiter() to allow a waiter to be set on an operation
which has alread been started.
Rewrote startDraining() and finishDrainingWithResult() in terms of
startOperation() and finishOperation().
Added a new OPERATING flag which is applied to all draining states.
Added four new OPERATING states, OPERATING, REBUILDING, and RESUMING,
and SUSPENDED_OPERATION. OPERATING is for operations which are still
considered part of normal operation. REBUILDING is a new drain state
used by read-only rebuild. RESUMING can only be entered from a
quiescent state. SUSPEND_OPERATION is for operations which can only be
performed in a suspended state and which leave the VDO suspended
afterward.
Split isOperatingNormally() into isNormal() and isOperating().
Added several new predicates, some on states, some on state codes.

.../blockMapInternals.h
.../blockMap.c
Reflected ActionManager changes.
Removed the 'state' field as it now lives in the ActionManager.

.../slabDepot.h
.../slabDepotInternals.h
.../slabDepot.c
Combined flushDepotSlabJournals(), suspendSlabSummary() and
saveSlabDepot() into a single drainSlabDepot() method.
Eliminated the now-unused 'saveRequested' field.
Renamed resumeSlabSummary() to resumeSlabDepot().
Reflected ActionManager changes.

.../slabJournal.c
Reflected AdminState and ActionManager changes.

.../readOnlyRebuild.c
.../vdoClose.c
.../vdoRecovery.c
.../vdoResize.c
Reflected SlabDepot changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>
Co-authored-by: bjohnston <bjohnsto@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>