139963 by corwin on 2019/07/10 (pair: sweettea) Merged 138558 to linux-vdo. 138558 by corwin on 2019/03/28 (pair: sclafani, msakai) Unified slabDepot load methods [VDO-4402].
This is another simplification prior to full suspend and resume
semantics for the slab depot and its components. This is the first of a
series of changes which will culminate in replacing SlabCompletion with
a less convoluted mechanism based on AdminStates and simple iteration
managed by the block allocators. This first step combines somewhat
similar code and uses the same type field as prepareToAllocate to
determine the type of load to do rather than having three different
functions to support the four flavors of load.

.../slabDepot.h
.../slabDepotInternals.h
.../slabDepot.c
Renamed the enumerated values of the SlabDepotLoadType to reflect the
type of load being done rather than the way the BlockAllocator should
interpret it. This makes more sense from the point of view of the
caller.
Added a fourth load type to represent a new slab depot.
Merged loadSlabDepotForRecovery() and loadSlabDepotForRebuild() into
loadSlabDepot(). Renamed the second parameter of loadSlabDepot()
and made it a SlabDepotLoadType to indicate which of the four flavors
of load is desired.
Added a loadWaiter field to hold the parent of the load operation
rather than storing it in the depot's SlabCompletion which is going
to go away.
Modernized loadSlabDepot() by using error handlers.

.../slabSummary.h
.../slabSummary.c
Modified loadSlabSummary() to take an error handler and use
prepareCompletion() (which didn't exist the last time this method
was touched). This is needed by the modernization of loadSlabDepot().

.../blockAllocator.c
.../readOnlyRebuild.c
.../vdoLoad.c
.../vdoRecovery.c
Reflected changes.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: sclafani <sclafani@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>