139964 by corwin on 2019/07/10 (pair: sweettea) Merged 138602 to linux-vdo. 138602 by corwin on 2019/04/01 (pair: msakai) Combined the flush and close passes of a slab when saving [VDO-4402].
Currently, when shutting down the VDO, for each slab, its journal is
flushed, the reference blocks are saved, the journal is closed, and
then the reference blocks are closed. However, because this save can
only be done once there are no active DataVIOs, there can be no slab
journal waiters when this process begins. As such, while saving the
dirty reference blocks can cause journal locks to be released, no
additional journal entries will be made as a result. Hence there is
no need for the double pass. This change is a simplification which
is another step in implementing suspend and resume.

.../slabCompletion.c
Modified saveSlabs() to only make one pass through the slab journal.

Co-authored-by: corwin <corwin@redhat.com>
Co-authored-by: sweettea <sweettea@redhat.com>
Co-authored-by: msakai <msakai@redhat.com>